update cart + order

+++ Backend/src/main/java/com/Hyperfume/Backend/service/impl/CartServiceImpl.java
@@ -36,21 +36,25 @@ public class CartServiceImpl implements CartService {
     UserRepository userRepository;
 
     public void addToCart(CartRequest request) {
-        PerfumeVariant variant = variantRepository
-                .findById(request.getVariantId())
-                .orElseThrow(() -> new AppException(ErrorCode.VARIANT_NOT_FOUND));
-
-        if (variant.getPerfume_stock_quantity() < request.getQuantity()) {
-            throw new AppException(ErrorCode.NOT_ENOUGH_STOCK_AVAILABLE);
-        }
-
-        Cart cart = cartMapper.toEntity(request);
 
         var context = SecurityContextHolder.getContext();
         String name = context.getAuthentication().getName();
 
         User user = userRepository.findByUsername(name).orElseThrow(() -> new AppException(ErrorCode.USER_NOT_EXISTED));
 
+        PerfumeVariant variant = variantRepository
+                .findById(request.getVariantId())
+                .orElseThrow(() -> new AppException(ErrorCode.VARIANT_NOT_FOUND));
+
+        if(cartRepository.existedByUserIdAndVariantId(user.getId(), variant.getId())){
+            throw new AppException(ErrorCode.VARIANT_ALREADY_IN_CART);
+        }
+
+        if (variant.getPerfume_stock_quantity() < request.getQuantity()) {
+            throw new AppException(ErrorCode.NOT_ENOUGH_STOCK_AVAILABLE);
+        }
+
+        Cart cart = cartMapper.toEntity(request);
         cart.setUser(user);
 
         cartRepository.save(cart);
@@ -82,6 +86,9 @@ public class CartServiceImpl implements CartService {
     }
 
     public void deleteCart(Integer cartId) {
+        if (!cartRepository.existsById(cartId)) {
+            throw new AppException(ErrorCode.CART_NOT_FOUND);
+        }
         cartRepository.deleteById(cartId);
     }
 }
